Adds Flasks
I have modified the Captains flask into the hipflask and I have created two new flasks, the commie flask which is a re-description of a madworld flask, and a russian flask which is a re-design of the detective flask. Hopefully this change will encourage more flask usage so when you go to the bar or are sharing a bottle among your mates, you divvy it up in these containers.